+++ package.json
@@ -31,7 +31,7 @@
     "react-dom": "^17.0.1",
     "react-helmet": "^6.1.0",
     "react-hotkeys-hook": "^2.4.0",
-    "react-i18next": "^11.7.4",
+    "react-i18next": "^11.8.3",
     "react-leaflet": "^3.0.2",
     "react-map-gl": "^5.2.10",
     "react-mapkit": "^0.6.5",
@@ -45,25 +45,26 @@
     "redux-thunk": "^2.3.0",
     "typescript": "^4.1.2",
     "use-debounce": "^5.1.0",
+    "use-throttle": "^0.0.3",
     "web-vitals": "^1.0.1"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.2",
-    "@testing-library/react-hooks": "^3.4.2",
+    "@testing-library/react-hooks": "^3.7.0",
     "@testing-library/user-event": "^12.5.0",
     "@types/deep-equal": "^1.0.1",
-    "@types/googlemaps": "^3.40.3",
-    "@types/jest": "^26.0.16",
-    "@types/node": "^14.14.10",
+    "@types/googlemaps": "^3.40.5",
+    "@types/jest": "^26.0.19",
+    "@types/node": "^14.14.12",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
     "@types/react-helmet": "^6.1.0",
     "@types/react-leaflet": "^2.5.2",
     "@types/react-map-gl": "^5.2.9",
-    "@types/react-redux": "^7.1.11",
+    "@types/react-redux": "^7.1.12",
     "cz-conventional-changelog": "3.3.0",
-    "husky": "^4.3.4",
+    "husky": "^4.3.5",
     "jest-environment-jsdom-sixteen": "^1.0.3",
     "lint-staged": "^10.5.3",
     "prettier": "^2.2.1",
