+++ public/translations/pl.json
@@ -8,7 +8,7 @@
   "search.label": "Szukaj",
   "search.noResults": "Nie znaleziono wyników",
   "settings.maps": "Mapy",
-  "settings.centeringMode.title": "Tryb Pozycjonowania",
+  "settings.centeringMode.title": "Pozycjonowanie",
   "settings.centeringMode.center": "Wyśrodkowanie",
   "settings.centeringMode.fill": "Wypełnienie",
   "settings.centeringMode.none": "Brak",
