+++ src/components/NavbarActions/MapSuppliers/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MapSuppliers';
