+++ src/enums/KeyboardShortcut.ts
@@ -2,12 +2,14 @@ export enum KeyboardShortcut {
   ToggleLock = 'l',
   ToggleTheme = 't',
   // TODO
-  // OpenHelp = 'h',
-  // Geolocation = 'g',
+  OpenHelp = 'h',
+  Geolocation = 'g',
   // FlipLeft = 'f+l',
   // FlipRight = 'f+r',
   // FlipTop = 'f+t',
   // FlipBottom = 'f+b',
   // CloseCurrent = 'x',
-  // CloseAll = 'x+a',
+  CloseAll = 'x+a',
+  HistoryUndo = 'u',
+  HistoryRedo = 'r',
 }
