+++ src/hooks/useLayout/index.ts
@@ -11,6 +11,7 @@ import {
 } from 'react-mosaic-component';
 import deepEqual from 'deep-equal';
 import { useTranslation } from 'react-i18next';
+// import { useThrottle } from 'use-throttle';
 
 import * as Actions from 'store/actions';
 
@@ -20,8 +21,12 @@ import { gridLayout, columnLayout, mosaicLayout } from 'constants/layouts';
 
 import type { Layout } from 'types/Layout';
 
+const layoutHistory: any[] = [];
+
 export const useLayout = () => {
   const { t } = useTranslation();
+
+  console.log(layoutHistory);
 
   const dispatch = useDispatch();
 
@@ -52,6 +57,8 @@ export const useLayout = () => {
   };
 
   const setActiveLayout = (layout: Layout) => {
+    layoutHistory.push(layout);
+
     dispatch(Actions.setActiveLayout(layout));
   };
